actualización backend render

+++ java/src/main/java/Emotia/java/repositorio/ITrepositorioUsuario.java
@@ -1,14 +1,19 @@
 package Emotia.java.repositorio;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.google.common.base.Optional;
 
 import Emotia.java.modelo.Usuarios;
 
-public interface ITrepositorioUsuario extends JpaRepository<Usuarios,Integer>{
+public interface ITrepositorioUsuario extends JpaRepository<Usuarios,Long>{
 
     Usuarios findByTelefono(String telefono);
 
     Usuarios findByNombre(String nombre);
+
+    //Usuarios findByIdusuarios(Long idUsuario);
 
 }
